refactor(Views v2) move List_Behavior in Traits
